src/final_login/pyproject.toml
[project]
name = "final_login"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "oddsummer56", email = "oddsummer56@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.32.1",
    "python-dotenv>=1.0.1",
    "pymongo>=4.9.2",
    "motor>=3.6.0",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
